:recycle: (PC-6484) remove useless use function
