+++ src/libs/geolocation/index.ts
@@ -1,4 +1,4 @@
 import CoordinatesView from './components/CoordinatesView'
 import { useGeolocation, GeolocationWrapper } from './GeolocationWrapper'
-import { useRequestGeolocPermission } from './useRequestGeolocPermission'
-export { CoordinatesView, useGeolocation, GeolocationWrapper, useRequestGeolocPermission }
+import { requestGeolocPermissionRoutine } from './requestGeolocPermissionRoutine'
+export { CoordinatesView, useGeolocation, GeolocationWrapper, requestGeolocPermissionRoutine }
