+++ src/libs/geolocation/requestGeolocPermissionRoutine.d.ts
@@ -1,0 +1,3 @@
+export const requestGeolocPermissionRoutine: (
+  setPermissionGranted: (granted: boolean) => void
+) => () => void
