--- src/libs/geolocation/useRequestGeolocPermission.d.ts
@@ -1,5 +1,0 @@
-export const useRequestGeolocPermission: (
-  setPermissionGranted: (granted: boolean) => void
-) => {
-  requestPermissionRoutine: () => void
-}
